+++ src/components/form/form-summary/__spec__.js
@@ -1,0 +1,68 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+
+import Icon from './../../icon';
+
+import FormSummary from './form-summary';
+
+describe('<FormSummary />', () => {
+  let block = '.carbon-form-summary',
+      wrapper;
+
+  describe("errors renders", () => {
+    beforeEach(() => wrapper = shallow( <FormSummary errors='1' /> ));
+
+    it("a block for errors if errors are provided", () => {
+      expect(wrapper.find(`${block}__error-summary`).length).toEqual(1);
+    });
+
+    it("an error icon", () => {
+      let icons = wrapper.find(Icon);
+      expect(icons.find('[type="error"]').length).toEqual(1);
+    });
+
+    it("the correct translation", () => {
+      expect(wrapper.text()).toContain('There is 1 error');
+    });
+  });
+
+  describe("warnings renders", () => {
+    beforeEach(() => wrapper = shallow( <FormSummary warnings='1' /> ));
+
+    it("a block for warnings if warnings are provided", () => {
+      expect(wrapper.find(`${block}__warning-summary`).length).toEqual(1);
+    });
+
+    it("an warning icon", () => {
+      let icons = wrapper.find(Icon);
+      expect(icons.find('[type="warning"]').length).toEqual(1);
+    });
+
+    it("the correct translation", () => {
+      expect(wrapper.text()).toContain('There is 1 warning');
+    });
+  });
+
+  describe("errors and warnings renders", () => {
+    beforeEach(() => wrapper = shallow( <FormSummary errors='1' warnings='1' /> ));
+
+    it("a block for errors if errors are provided", () => {
+      expect(wrapper.find(`${block}__error-summary`).length).toEqual(1);
+    });
+
+    it("a block for warnings if warnings are provided", () => {
+      expect(wrapper.find(`${block}__warning-summary`).length).toEqual(1);
+    });
+
+    it("both icons", () => {
+      let icons = wrapper.find(Icon);
+      ['error','warning'].forEach(type => expect(icons.find(`[type='${type}']`).length).toEqual(1));
+    });
+
+    it("the correct translation", () => {
+      let text = wrapper.text();
+      expect(text).toContain('There is 1 error');
+      expect(text).toContain('and 1 warning');
+    });
+  });
+});